QEXPAND is a Quick-Edit tool only. It expands COPY, ++INCLUDE and -INC statements within source. The member referenced by one of those keywords is inserted as NOTE lines.  APIALPKG is an API program to be used with the PACKAGE tool.
QEXPAND is a Quick-Edit tool only. It expands COPY, ++INCLUDE and -INC statements within source. The member referenced by one of those keywords is inserted as NOTE lines.

APIALPKG is an API program to be used with the PACKAGE tool.  The user can enter "TSO PACKAGE" on a command line, and move the cursor to an existing package name, before pressing 'Enter'. The Package tool will then create a new package and copy the contents whose name appears on the screen.